feat: LottoGenerator 인터페이스 생성 후 테스트를 위해 수동으로 로또 생성할 수 있는 기능 추가
