+++ src/main/java/lotto/Application.java
@@ -13,10 +13,10 @@ public class Application {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
 
+        LottoMachine lottoMachine = LottoMachine.getInstance();
         Money money = new Money(InputView.takeMoneyInput(scanner));
-        LottoGenerator lottoGenerator = new LottoGenerator();
 
-        Lottos lottos = new Lottos(lottoGenerator.createLottos(money.calculateAffordableNumberOfLotto()));
+        Lottos lottos = new Lottos(lottoMachine.createLottos(new AutomaticLottoGenerator(), money.calculateAffordableNumberOfLotto()));
         OutputView.showLottos(lottos);
 
         WinningLotto winningLotto = new WinningLotto(
