+++ src/test/java/lotto/domain/ManualLottoGenerator.java
@@ -1,0 +1,29 @@
+package lotto.domain;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class ManualLottoGenerator implements LottoGenerator {
+    private final List<Integer> numbers;
+
+    public ManualLottoGenerator(String numbers) {
+        this.numbers = Arrays.stream(numbers.split(", "))
+                .map(Integer::parseInt)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public Lotto createLotto() {
+        return new Lotto(createLottoNumbers());
+    }
+
+    private List<LottoNumber> createLottoNumbers() {
+        List<LottoNumber> lottoNumbers = new ArrayList<>();
+        for (Integer number : numbers) {
+            lottoNumbers.add(new LottoNumber(number));
+        }
+        return lottoNumbers;
+    }
+}
